# backend/openapi.yml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Testing Nirvana API
servers:
  - url: api
components:
  responses:
    GenericErrorResponse:
      description: The request failed.
      content:
        application/json:
          schema:
            properties:
              message:
                type: string
              code:
                type: string
              errors:
                type: array
                items:
                  type: string
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phoneNumber:
          type:
            - string
            - "null"
        securityLevel:
          type: string
          enum:
            - FULL_ACCESS
            - LIMITED
            - READ_ONLY
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
        school:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
          required:
            - id
            - name
        role:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
          required:
            - id
            - name
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - firstName
        - lastName
        - email
        - phoneNumber
        - securityLevel
        - status
        - school
        - role
        - createdAt
        - updatedAt
    Pagination:
      type: object
      properties:
        pages:
          type: number
        total:
          type: number
      required:
        - pages
        - total
...
paths:
  /users:
    get:
      description: Get list of users
      parameters:
        - schema:
            type: string
          required: false
          name: name
          in: query
        - schema:
            type: string
          required: false
          name: school
          in: query
        - schema:
            type: string
            enum:
              - ADMIN
              - TEACHER
              - THERAPIST
              - PROVIDER
              - SUPERVISOR
          required: false
          name: role
          in: query
        - schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
          required: false
          name: status
          in: query
        - schema:
            type: string
            enum:
              - name
              - school
              - role
              - status
            default: name
          required: false
          name: sortBy
          in: query
        - schema:
            type: string
            enum:
              - asc
              - desc
            default: asc
          required: false
          name: sortOrder
          in: query
        - schema:
            type: number
            minimum: 10
            maximum: 300
            default: 20
          required: false
          name: perPage
          in: query
        - schema:
            type: number
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
      responses:
        "200":
          description: Object with data and pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
                required:
                  - data
                  - pagination
        "400": &a1
          $ref: "#/components/responses/GenericErrorResponse"
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    post:
      description: Create a user
      parameters:
        - schema:
            type:
              - boolean
              - "null"
            default: false
          required: false
          name: sendInvite
          in: query
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                securityLevel:
                  type: string
                  enum:
                    - FULL_ACCESS
                    - LIMITED
                    - READ_ONLY
                schoolId:
                  type: string
                  format: uuid
                roleId:
                  type: string
                  format: uuid
                phoneNumber:
                  type:
                    - string
                    - "null"
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
              required:
                - firstName
                - lastName
                - email
                - securityLevel
                - schoolId
                - roleId
                - phoneNumber
                - status
      responses:
        "204":
          description: The user has been created (and invited, if chosen) successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
  /users/{id}:
    get:
      description: Get a user
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Object with data containing user information.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/User"
                required:
                  - data
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    patch:
      description: Update a user
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                securityLevel:
                  type: string
                  enum:
                    - FULL_ACCESS
                    - LIMITED
                    - READ_ONLY
                schoolId:
                  type: string
                  format: uuid
                roleId:
                  type: string
                  format: uuid
                phoneNumber:
                  type:
                    - string
                    - "null"
                status:
                  type: string
                  enum:
                    - ACTIVE
                    - INACTIVE
      responses:
        "204":
          description: The user has been updated successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
    delete:
      description: Delete a user
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "204":
          description: The user has been deleted successfully.
        "400": *a1
        "401": *a1
        "403": *a1
        "404": *a1
        "500": *a1
...